tools/era_import_single/requirements.txt
pyoxigraph>=0.5.3
psycopg[binary]>=3.1.0
python-dotenv>=1.0.0
